+++ packages/thu-info-lib/src/models/schedule/calendar.ts
@@ -1,5 +1,8 @@
-export interface CalendarData {
+export interface Semester {
     firstDay: string;   // yyyy-MM-dd
     semesterId: string;
     weekCount: number;
+}
+export interface CalendarData extends Semester {
+    nextSemesterList: Semester[];
 }